refactor(user): convert to module
- create the user module, move controller, services, types therein
- eliminate controller and service interfaces
