--- src/server/controllers/types.ts
@@ -1,5 +1,0 @@
-export type GoUploadedFile = {
-  data: Buffer
-  mimetype: string
-  name: string
-}
