+++ src/server/inversify.config.ts
@@ -41,8 +41,8 @@ import { SentryController } from './modules/sentry/SentryController'
 import { LoginController } from './controllers/LoginController'
 import { AuthService } from './services/AuthService'
 import { LogoutController } from './controllers/LogoutController'
-import { UrlManagementService } from './services/UrlManagementService'
-import { UserController } from './controllers/UserController'
+import { UrlManagementService } from './modules/user/services'
+import { UserController } from './modules/user'
 import { QrCodeService } from './services/QrCodeService'
 import { DirectoryController } from './controllers/DirectoryController'
 import { DirectorySearchService } from './services/DirectorySearchService'
