+++ src/server/modules/user/services/index.ts
@@ -1,0 +1,3 @@
+export { UrlManagementService } from './UrlManagementService'
+
+export { default } from './UrlManagementService'
