--- src/server/services/types.ts
@@ -1,8 +1,0 @@
-import { StorableUrl } from '../repositories/types'
-import { GoUploadedFile } from '../controllers/types'
-
-export type UpdateUrlOptions = Partial<
-  Pick<StorableUrl, 'state' | 'longUrl' | 'description' | 'contactEmail'> & {
-    file: GoUploadedFile
-  }
->
